refactor: noticeView 분리
